Auto-MPD/requirements.txt
nibabel==3.2.1numpy==1.21.4opencv-python==4.5.4.60packaging==21.3Pillow==8.4.0pydicom==2.2.2pynrrd==0.4.2pyparsing==3.0.6SimpleITK==2.1.1
matplotlib==3.5.1 
scikit-image==0.19.1 
pandas==1.3.5
connected-components-3d==3.6.0 
scipy==1.6.2 
pyyaml==6.0